requirements.txt
Babel==2.6.0
dj-database-url==0.5.0
Django==2.1.3
django-bootstrap4==0.0.7
django-heroku==0.3.1
django-multiselectfield==0.1.8
django-phonenumber-field==2.1.0
django-tinymce4-lite==1.7.3
django-widget-tweaks==1.4.3
factory-boy==2.11.1
Faker==1.0.0
gunicorn==19.9.0
jsmin==2.2.2
phonenumbers==8.9.16
phonenumberslite==8.9.16
psycopg2==2.7.6.1
python-dateutil==2.7.5
python-http-client==3.1.0
pytz==2018.7
raven==6.9.0
sendgrid==5.6.0
six==1.11.0
text-unidecode==1.2
urllib3==1.24.1
whitenoise==4.1.1
